+++ dagfactory/__init__.py
@@ -1,2 +1,2 @@
 """Modules and methods to export for easier access"""
-from .dagfactory import DagFactory
+from .dagfactory import DagFactory, load_yaml_dags