+++ CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.1] - 2023-01-21
+### Changed
+- Changed imports to support Kubernetes Provider > 5.0.0
+
+## [0.17.0] - 2022-12-06
+### Added
+- Adds `sla_secs` in `default_args` to convert seconds to `timedelta` obj
+### Changed
+- Changed deprecated imports to support Airflow 2.5
+- Removed support for Python 3.6
+
+## [0.16.0] - 2022-11-13
+### Added
+- Function to scan recursively for YAML DAGs 
+### Changed
+- Changed deprecated imports to support Airflow 2.4+
+
 ## [0.15.0] - 2022-09-09
 ### Added
 - Support for string concatenation of variables in YAML
@@ -148,7 +165,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-11-20
 - Initial release
 
-[Unreleased]: https://github.com/ajbosco/dag-factory/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/ajbosco/dag-factory/compare/v0.17.1...HEAD
+[0.17.1]: https://github.com/ajbosco/dag-factory/compare/v0.17.0...v0.17.1
+[0.17.0]: https://github.com/ajbosco/dag-factory/compare/v0.16.0...v0.17.0
+[0.16.0]: https://github.com/ajbosco/dag-factory/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/ajbosco/dag-factory/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/ajbosco/dag-factory/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/ajbosco/dag-factory/compare/v0.12.0...v0.13.0